move HttpResponseStatus to the feature specific package
